requirements.txt
django==4.2.4
django-cors-headers==4.2.0
djangorestframework  #==3.14.0
djangorestframework-simplejwt #==5.3.0
django-filter==23.2
requests==2.31.0
python-dotenv==1.0.0
mysqlclient
django-redis